backend/requirements.txt
# Core dependencies
fastapi==0.115.10
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.10.6
pydantic-settings==2.3.4

# Auth and Security
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ML / Search (lightweight — no model downloads)
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
qdrant-client>=1.7.0
google-genai>=1.0.0
langchain-text-splitters>=0.2.0

# Document Processing
python-multipart==0.0.6
python-docx==1.0.1
pypdf2==3.0.1
python-slugify==8.0.1
chardet==5.2.0

# Web Scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
